+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Don't call `refresh!` on non-refreshable `Credentials` when retrying errors (#2685).
+
 3.130.0 (2022-03-11)
 ------------------